Extended teh capablities and data format of ultraScurve
